Bit Magic 1
Given an integer t - the number of test cases. For each test case, read an integer n and answer the following 6 queries in the given order in a new line -:

Output the 64 bit binary representation of number n.
Output the position of the most significant bit ( MSB ) or the leftmost set bit of n ( 0- indexed ). If n=0, output -1.
Output the position of the rightmost set bit of n ( 0- indexed ). If n=0, output -1.
Output 1 if n  is a power of 2 i.e n==2X , x > 0, else output 0.
Output the biggest power of 2, that is the divisor of n. i.e biggest 2K such that n%2K == 0, k ≥ 0. If n=0, output -1. 
Output the smallest power of 2, which is not smaller than n, i.e smallest 2K such that 2K ≥ n, k>0.