Add files via upload
Description
Your city is having people infected with a virus. The city in which you live is represented as a grid consisting of n rows and m columns. Cells containing 2 are the cells where the people infected with the virus are present and the cells having 1 are the cells having people not yet infected with the virus. There are certain empty cells which are represented by 0. The infected people in a unit time can infect all their adjacent cells, i.e, if they are present at cell [i, j] they can infect cells [i-1, j], [i+1, j], [i, j-1] and [i, j+1]. The virus cannot pass through empty cells. Your task is to print the minimum time in which all the people are infected with the virus. If the virus cannot infect everyone, print -1.

Input Format
The first line contains two integers n and m  — the number of rows and columns, respectively.
The following n lines contain m integers each, the j-th element in the i-th line is the number written in the j-th cell of the i-th row.

Output Format
Print the minimum time in which everyone can be infected or -1 if everyone cannot be infected.

Constraints
1 <= n <= 1000
1 <= m <= 1000
Sample Input 1
3 3
0 1 2
0 1 2
1 1 2
Sample Output 1
2
Sample Input 2
3 3
0 1 2
0 1 2
1 0 2
Sample Output 2
-1