Add files via upload
Here's the corrected problem statement:

---

## Description
The Fibonacci Sequence is a series of numbers, in which the \(i\)-th number is given by the recurrence relation:
\[ F(i) = F(i-1) + F(i-2) \]
where \( F(0) = 0 \) and \( F(1) = 1 \).

Given an integer \( N \, (\ge 2) \), you need to find its index in the Fibonacci Sequence or print \(-1\) if it is not a part of the sequence.

## Input Format
The first line of input contains one integer \( T \) - the number of test cases. Then \( T \) test cases follow.

The only line of each test case contains an integer \( N \).

## Output Format
For each test case, print on a new line the index of \( N \) in the Fibonacci Sequence, or \(-1\) if it is not a part of the sequence.

## Constraints
1. \( 1 \le T \le 10 \)
2. \( 2 \le N \le 10^{15} \)

## Sample Input 1
```
2
13
34
```

## Sample Output 1
```
7
9
```

## Sample Input 2
```
2
21
10
```

## Sample Output 2
```
8
-1
```

---

You can now copy and paste this problem statement.