Add files via upload
Description
You are given an integer N. You have to tell whether the number is a perfect square/cube according to the given query. Do not use inbuilt sqrt() and cube() function. As sqrt(4) can be equal to 1.9999999 and int(1.999999). = 1. Due to precision issue do not use sqrt() if you are dealing with integers.

Input Format
The first line of the input contains one integer T - the number of test cases. Then T test cases follow.

The first line of each test case contains two space-separated integers Q, N - Q=1 denotes that you have to check if N is a perfect square and Q=2 denotes that you have to check if N is a perfect cube.

Output Format
For each test case, print “YES” if the number is a perfect square/cube and “NO” if the number is not a perfect square/cube, according to the given query.

Constraints
1 ≤ T ≤ 100

1 ≤ N ≤ 107

Sample Input 1
4
1 64
2 64
2 9
1 5
Sample Output 1
YES
YES
NO
NO