Add files via upload
Description
You are given an array A of N integers. Find the number of pairs (i, j) such that i < j and Ai + Aj is a power of 2.

Input Format
The first line of the input contains one integer T - the number of test cases. Then T test cases follow.

The first line of each test case contains one integer N - the length of the array. 

The second line of each test case contains N space-separated integers.

Output Format
For each test case, print the number of pairs (i, j) such that i < j and Ai + Aj is a power of 2.

Constraints
1≤ T ≤ 105

1≤ N ≤ 105

1≤ Ai ≤ 230

It is guaranteed that the sum of N over all test cases does not exceed 105.

Sample Input 1
3
5
1 7 2 3 1
3
1 1 1
3
2 6 14
Sample Output 1
5
3
2